test: Test serializers for WSGI apps
