+++ test/test_serialization.py
@@ -1,0 +1,79 @@
+import csv
+from io import StringIO
+
+import pytest
+from hypothesis import given, settings
+
+import schemathesis
+
+
+def to_csv(data):
+    if not data:
+        return ""
+    output = StringIO()
+    field_names = sorted(data[0].keys())
+    writer = csv.DictWriter(output, field_names)
+    writer.writeheader()
+    writer.writerows(data)
+    return output.getvalue()
+
+
+@pytest.fixture
+def csv_serializer():
+    @schemathesis.serializers.register("text/csv")
+    class CSVSerializer:
+        def as_requests(self, context, value):
+            return {"data": to_csv(value)}
+
+        def as_werkzeug(self, context, value):
+            return {"data": to_csv(value)}
+
+    yield
+
+    schemathesis.serializers.unregister("text/csv")
+
+
+@pytest.fixture(params=["aiohttp", "flask"])
+def api_schema(request, openapi_version):
+    if request.param == "aiohttp":
+        schema_url = request.getfixturevalue("schema_url")
+        return schemathesis.from_uri(schema_url)
+    app = request.getfixturevalue("flask_app")
+    return schemathesis.from_wsgi("/schema.yaml", app=app)
+
+
+@pytest.mark.hypothesis_nested
+@pytest.mark.endpoints("csv_payload")
+@pytest.mark.usefixtures("csv_serializer")
+def test_text_csv(api_schema):
+    # When API expects `text/csv`
+    # And the user registers a custom serializer for it
+
+    @given(case=api_schema["/csv"]["POST"].as_strategy())
+    @settings(max_examples=5)
+    def test(case):
+        if case.app is not None:
+            response = case.call_wsgi()
+        else:
+            response = case.call()
+        # Then this serializer should be used
+        case.validate_response(response)
+        # And data should be successfully sent to the API as CSV
+        if case.app is not None:
+            data = response.json
+        else:
+            data = response.json()
+        assert data == case.body
+
+    test()
+
+
+def test_register_incomplete_serializer():
+    # When register a new serializer without a required method
+    # Then you'll have a TypeError
+    with pytest.raises(TypeError, match="`CSVSerializer` is not a valid serializer."):
+
+        @schemathesis.serializers.register("text/csv")
+        class CSVSerializer:
+            def as_requests(self, context, value):
+                return {}
